Fix issues with game speed due to game Clock
The clock did not update every frame causing large elapsed times when
resuming the game after a pause. The clock tried to implement an own
kind of vertical sync system and bypassing the one of SFML that way.

The ball's and paddle's speed have been adjusted for the new changes
to the clock.